server/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-cors
flask-jwt-extended